LDn-binning has been added
Note: this is a beta version 0.63 which was used for 'Linkage disequilibrium cluster-based approach for association mapping with tightly linked genome-wide data' in review for Mol Ecol Res. 

It works well with code provided with this publication (available from dryad), but has not properly been tested with other data. Documentation is also not yet complete, I aim to add a tutorial specifically for the LDn-binning approach, in the near future. I also need to make sure existing users of LDna can continue using it as before.