Update LDnClustering.R
Bugfix that affects how many PCs are extracted. This mainly affects LD clustering when low LD thresholds are used. This bug resulted in that the first PC and the PC that cumulatively explained >PC_threshold of the total genetic variation was extracted (not necessarily the first and second PC). This resulted in maximum two PCs being extracted, and apart from the first PC, not necessarily the two PCs that explained most of the variation. With high LD thresholds, the loci already already are highly correlated and this error was very rare.